Moved QrCodeDao to repository, switched to Short, other misc.
Moved the QrCodeDao from list based to now being JPA repository based.
Switched all UUID or Long keys to Shorts. Highly doubt the number of
users or qrcodes will go above 32k. The switch was to reduce potential
memory usage, especially since hosting isn't free.

+++ src/main/java/com/example/demo/dao/PersonDao.java
@@ -16,9 +16,9 @@ public interface PersonDao {
 
 	void insertPerson(Person person);
 	
-	int deletePersonById(Long id);
+	int deletePersonById(Short id);
 	
 	Iterable<Person> getAllPersonsIterable();
 	
-	Optional<Person> getPersonById(Long id);
+	Optional<Person> getPersonById(Short id);
 }
