+++ src/main/java/com/example/demo/repositories/PersonRepository.java
@@ -8,7 +8,7 @@ import com.example.demo.model.Person;
 // Why Long instead of UUID? (space) https://jivimberg.io/blog/2018/11/05/using-uuid-on-spring-data-jpa-entities/
 // DAO and JPA? https://stackoverflow.com/questions/15554826/new-to-java-whats-jpa-and-dao
 // What is JPA? https://www.infoworld.com/article/3379043/what-is-jpa-introduction-to-the-java-persistence-api.html
-public interface PersonRepository extends CrudRepository<Person, Long> {
+public interface PersonRepository extends CrudRepository<Person, Short> {
 	// This doesn't have anything in it, because CrudRepository<Person, Long> has a bunch of what we need already. 
 	// https://en.wikipedia.org/wiki/Convention_over_configuration
 }
